Initial commit
Starting with express generator, I added some CRUD on top of thinky (rethinkdb)
for two models: plan and step. Added a linter file, however the generated files
from the express generator are causing a bunch of linter errors at this point.
User routes are not used but was part of the express generator.