Preserve the order of a plan's steps. Tested!
Makes sure that we keep track of the order that the steps are in. Since the
steps are saved independently of each other then an order needs to be set
somewhere.

While the order could be set on the plan object, I felt it was best to keep the
plan from knowing about its steps at that level. So I added an `order` property
to the step. The only problem is that this order needs to be updated whenever
a step is added.

I'm just remembering now that the order is not updated on deletes. This could
result in some weirdness.