Preserving step detail item order.
By converting the step details to an array of objects instead of just one
object with keys, I was able to keep the order. Beforehand, the keys were
being sorted alphabetically.

The `order` param is not explicitely needed at this point. The transition to
an array seemed to solve the problem.