Password Reset
This commit enables password resets. Upon a reset request, it creates
a JSON web token with that email address and sends it off using
Mailgun. The user can then demonstrates ownership of the email account
by opening the link which includes the JSON web token. The token is
decoded and the email address is pulled. We then present the user with
a password form and update the account that belongs to the email
address.

JSON web tokens is valuable asset since we don't need to manage any
password reset requests with tokens in our database.